# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Freemily",
    "description": "Freemily 백엔드 서버를 nodejs로 재구현한 프로젝트입니다.",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/groups/ingredients/{group_id}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ingredientCount": {
                  "example": "any"
                },
                "ingredientCategory": {
                  "example": "any"
                },
                "ingredientSaveType": {
                  "example": "any"
                },
                "ingredientPurchaseDate": {
                  "example": "any"
                },
                "ingredientExpirationDate": {
                  "example": "any"
                },
                "ingredientName": {
                  "example": "any"
                },
                "ingredientMemo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "saveType",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/groups/ingredients/{group_id}/details/{ingredient_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "ingredient_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/groups/ingredients/{group_id}/{ingredient_id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "ingredient_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ingredientName": {
                  "example": "any"
                },
                "ingredientSaveType": {
                  "example": "any"
                },
                "ingredientPurchaseDate": {
                  "example": "any"
                },
                "ingredientExpirationDate": {
                  "example": "any"
                },
                "ingredientCategory": {
                  "example": "any"
                },
                "ingredientCount": {
                  "example": "any"
                },
                "ingredientMemo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}